@@ -6,7 +6,7 @@ namespace Conrec.Domain.Entities
         public string PhoneNumber { get; set; }
 
         #region Links
-        public User User { get; set; }
+        public virtual User User { get; set; }
         public int UserId { get; set; }
         #endregion
     }
