@@ -3,7 +3,7 @@ namespace Conrec.Domain.Entities
     public class Country : RefType
     {
         #region Links
-        public Employee Employee { get; set; }
+        public virtual Employee Employee { get; set; }
         #endregion
     }
 }
