@@ -10,7 +10,7 @@ namespace Conrec.Domain.Entities
         public DateTimeOffset CreateDate { get; set; }
 
         #region Links
-        public User User { get; set; }
+        public virtual User User { get; set; }
         public int UserId { get; set; }
         #endregion
     }
