@@ -3,7 +3,7 @@ namespace Conrec.Domain.Entities
     public class JobRole : RefType
     {
         #region Links
-        public Employer Employer { get; set; }
+        public virtual Employer Employer { get; set; }
         #endregion
     }
 }
