@@ -3,7 +3,7 @@ namespace Conrec.Domain.Entities
     public class SalaryPayment : RefType
     {
         #region Links
-        public Team Team { get; set; }
+        public virtual Team Team { get; set; }
         #endregion
     }
 }
