@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 
 namespace Conrec.Domain.Entities
 {
@@ -6,16 +7,17 @@ namespace Conrec.Domain.Entities
     {
         public int Id { get; set; }
         public bool IsBreakPaid { get; set; }
+        public DateTimeOffset BreakStart { get; set; }
+        public DateTimeOffset BreakEnd { get; set; }
+        public DateTimeOffset DayStart { get; set; }
+        public DateTimeOffset DayEnd { get; set; }
         public double TotalHoursWorked { get; set; }
-        public DateTimeOffset BreakTime { get; set; }
-        public DateTimeOffset StartTime { get; set; }
-        public DateTimeOffset EndTime { get; set; }
 
         #region Links
         public int WorkDayId { get; set; }
-        public WorkDay WorkDay { get; set; }
-        public int ProjectId { get; set; }
-        public Project Project { get; set; }
+        public virtual WorkDay WorkDay { get; set; }
+        public virtual ICollection<ProjectSchedule> ProjectSchedules { get; set; }
+
         #endregion
     }
 }
