@@ -14,19 +14,20 @@ namespace Conrec.Domain.Entities
         public int Id { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
+        public string Password { get; set; }
         public string Email { get; set; }
         public DateTimeOffset? RegisterDate { get; set; }
 
         #region Links
-        public Employee Employee { get; set; }
-        public Employer Employer { get; set; }
+        public virtual Employee Employee { get; set; }
+        public virtual Employer Employer { get; set; }
         public int UserRolesId { get; set; }
-        public UserRole UserRole { get; set; }
+        public virtual UserRole UserRole { get; set; }
         public int RegionId { get; set; }
-        public Region Region { get; set; }
-        public ICollection<Contact> Contacts { get; private set; }
-        public ICollection<Bank> Banks { get; private set; }
-        public ICollection<Notification> Notifications { get; private set; }
+        public virtual Region Region { get; set; }
+        public virtual ICollection<Contact> Contacts { get; private set; }
+        public virtual ICollection<Bank> Banks { get; private set; }
+        public virtual ICollection<Notification> Notifications { get; private set; }
         #endregion
     }
 }
