@@ -3,7 +3,7 @@ namespace Conrec.Domain.Entities
     public class UserRole : RefType
     {
         #region Links
-        public User User { get; set; }
+        public virtual User User { get; set; }
         #endregion
     }
 }
