@@ -3,7 +3,7 @@ namespace Conrec.Domain.Entities
     public class WorkDay : RefType
     {
         #region Links
-        public Schedule Schedule { get; set; }
+        public virtual Schedule Schedule { get; set; }
         #endregion
     }
 }
